Fix possible duplicate home menu entries
If the home menu is ordered when a pending home menu request exist, we
may see duplicate home menu entries. This may happe e.g. on network
issues.

To avoid this we perform the clear and add of the home menu in one
operation when all home menu items hve been received.
